Split exporter module classs
